chore: misc changes per Anthony review
