+++ tests/rule_based_profiler/config/test_base.py
@@ -1,6 +1,9 @@
+from typing import Dict
+
 import pytest
 from ruamel.yaml.comments import CommentedMap
 
+from great_expectations.core.batch import BatchRequest
 from great_expectations.marshmallow__shade.exceptions import ValidationError
 from great_expectations.rule_based_profiler.config import (
     DomainBuilderConfig,
@@ -15,6 +18,7 @@ from great_expectations.rule_based_profiler.config import (
     RuleConfig,
     RuleConfigSchema,
 )
+from great_expectations.rule_based_profiler.rule_based_profiler import RuleBasedProfiler
 
 
 def test_not_null_schema_raises_error_with_improperly_implemented_subclass():
@@ -279,3 +283,69 @@ def test_rule_based_profiler_from_commented_map():
     commented_map = CommentedMap(data)
     config = RuleBasedProfilerConfig.from_commented_map(commented_map)
     assert all(hasattr(config, k) for k in data)
+
+
+def test_resolve_config_using_acceptable_arguments(
+    profiler_with_placeholder_args: RuleBasedProfiler,
+) -> None:
+    config: dict = RuleBasedProfilerConfig.resolve_config_using_acceptable_arguments(
+        profiler=profiler_with_placeholder_args
+    )
+
+    # Ensure we have expected keys while also removing unnecessary ones
+    assert all(
+        attr in config
+        for attr in ("name", "config_version", "rules", "variable_count", "rule_count")
+    )
+    assert all(
+        attr not in config for attr in ("class_name", "module_name", "variables")
+    )
+
+    assert config["variable_count"] == 1 and config["rule_count"] == 1
+
+
+def test_resolve_config_using_acceptable_arguments_with_runtime_overrides(
+    profiler_with_placeholder_args: RuleBasedProfiler,
+) -> None:
+    rule_name: str = "my_rule"
+    assert all(rule.name != rule_name for rule in profiler_with_placeholder_args.rules)
+
+    rules: Dict[str, dict] = {rule_name: {"foo": "bar"}}
+    config: dict = RuleBasedProfilerConfig.resolve_config_using_acceptable_arguments(
+        profiler=profiler_with_placeholder_args, rules=rules
+    )
+
+    assert len(config["rules"]) == 1 and rule_name in config["rules"]
+
+
+def test_resolve_config_using_acceptable_arguments_with_runtime_overrides_with_batch_requests(
+    profiler_with_placeholder_args: RuleBasedProfiler, usage_stats_profiler_config: dict
+) -> None:
+    datasource_name = "my_datasource"
+    data_connector_name = "my_basic_data_connector"
+    data_asset_name = "my_data_asset"
+
+    batch_request: BatchRequest = BatchRequest(
+        datasource_name=datasource_name,
+        data_connector_name=data_connector_name,
+        data_asset_name=data_asset_name,
+    )
+
+    # Add batch requests to fixture before running method
+    rules: Dict[str, dict] = usage_stats_profiler_config["rules"]
+    rules["rule_1"]["domain_builder"]["batch_request"] = batch_request
+
+    config: dict = RuleBasedProfilerConfig.resolve_config_using_acceptable_arguments(
+        profiler=profiler_with_placeholder_args, rules=rules
+    )
+
+    assert all(
+        attr in config
+        for attr in ("name", "config_version", "rules", "variable_count", "rule_count")
+    )
+
+    domain_builder: dict = config["rules"]["rule_1"]["domain_builder"]
+    converted_batch_request: dict = domain_builder["batch_request"]
+    assert converted_batch_request["datasource_name"] == datasource_name
+    assert converted_batch_request["data_connector_name"] == data_connector_name
+    assert converted_batch_request["data_asset_name"] == data_asset_name
